+++ package.json
@@ -7,11 +7,16 @@
   "dependencies": {
     "react": "16.12.0",
     "react-dom": "16.12.0",
-    "react-scripts": "3.0.1",
+    "react-scripts": "3.3.0",
     "roughjs": "3.1.0"
   },
   "devDependencies": {
-    "typescript": "3.3.3"
+    "@types/react": "16.9.17",
+    "@types/react-dom": "16.9.4",
+    "husky": "3.1.0",
+    "lint-staged": "9.5.0",
+    "prettier": "1.19.1",
+    "typescript": "3.7.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -24,5 +29,22 @@
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ]
+  ],
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.{js,css,json,md,ts,tsx}": [
+      "prettier --write",
+      "git add"
+    ],
+    "*.{js,ts,tsx}": [
+      "eslint"
+    ]
+  }
 }
