+++ src/types/rough.d.ts
@@ -1,0 +1,1 @@
+declare module "roughjs/dist/rough.umd.js";
